+++ FluidSimulator.cpp
@@ -158,7 +158,7 @@ void FluidSimulator::fillKernelLookupTables()
 #pragma omp parallel for
     for (int i = 0; i < lookupTableSize; ++i)
     {
-        float squaredDistance = i * 1e-4;
+        float squaredDistance = i * 1e-4f;
         poly6LookupTable[i] = poly6(squaredDistance, h);
     }
 }
@@ -187,7 +187,7 @@ float FluidSimulator::calcDensity(const glm::vec3& pos) const
                         float sqLen = glm::length2(pos - r[j]);
                         if (sqLen < h*h)
                         {
-                            int lookupIndex = 1e4 * glm::length2(pos - r[j]);
+                            int lookupIndex = (int)(1e4f * glm::length2(pos - r[j]));
                             rho += m * poly6LookupTable[lookupIndex];
 
                             // Non-lookup table version:
